This is how I want the loop to be. At least for now. I'll add short circuiting in the iterator later.
Now...the thing I will have to do is make the damn optimizer.

I am going to have to iterate over the network and get the weights. The way I will do that is by generalizing the init's traversal. For the sake of the future, I am also thinking that it might be worth it to leave weights as a function. I mean just a `const ()`. That will allow me to compose initialized networks with unitialized ones.

Let me make some changes then.