I am dithering back and forth right now.
It is interesting. The exercise of generalization is something that really tests the understanding of the subject and has the effect of solidifying the architecture.

It is kind of a waste of time from a certain perspective, but let me see if I can get it done regardless.

It is important that I maximize my understanding rather than just go with the first thing. The long term will pay me for it.