I am almost ready to start the day.
The truth is even though I caught up yesterday, I am not at all sure what I should be doing today.

But one thing that stands out to me is just how ugly learning9 is.

The new iterators are a step in the right direction, but I haven't gotten all the way there yet.

Yeah, I am absolutely sure that my approach towards composing layers is correct, but that is not all.

I need to optimize the ergonomics of actually running the net to their limit.

Right now as it is so small, I am learning towards removing fold.