Here is the new design for the fold and iterator.
The two fold functions have a lot of shared structure, but I know from experience to not try to optimize in these spots because the gaps are too narrow to slip through.

The region handling in iter makes it look ugly, but all in all, I am proud of the current design.

The code is currently broken. I'll make it work after breakfast so let me stop here for a bit.