With this everything is cleaned up. Maybe I am missing some things for adaptive learning rates, but screw that.
I'll think about it when I need it. No doubt, I will extend the functions in order to compose with the recurrent parts, but right now that is not important.

I feel like taking a slight breather here.

...Really, I did a lot of work over the last few days to make everything fit.

If there is a better way than this to make a NN library, I'd sure like to see it.

Here everything is purposeful and there is zero wasted code. The fundamentals are in place for the library to grow.

Well written code is bliss. It is something one can truly own. It is the wings on my back. Poorly written code on the other hand is liability. It is an anchor weighting me down.

Right now the amount of poorly written code in the library amounts to zero.

I will stop here just to enjoy the moment. Once I get back I will start making plans for the next step.