I will take responsibility for this.
Unlike usual, I do not at all feel like programming today. I've made my determination and mostly feel like arguing on Reddit today.

At any rate, today I just feel like talking and socializing so let me use this commit for that.

First of all, I think it is time for a change regarding how I do programming.

I've been doing this for three years and I've proved my commitment. It is not like 2015 when I was starting out when I doubted my determination and set out to prove it at all costs.

In 2017 of all things I actually had to see a doctor because I've been pushing myself so hard the stress started to feel like ulcers. I mean it literally started manifesting as physical pain.

Yesterday, I finished my round of programming at 4:30pm. That is when I made a commit, was out of inspiration and then decided to watch a lecture rationalizing it as work.

The end result it that I wasted my time, was bored out of my skull the whole time and when I finished the day at 7pm my stress was through the roof.

Now today, my inspiration was refilled as it usual is overnight. But instead of programming I decided it was essay time.

I...just what am I doing?

Does it really make sense for me to continue programming in a routine every day like it was a job or maybe I should get rid of my restrictions and do only what I feel like doing every day?

Currently the library is 1.9k lines of code. Spread over 3 months, that comes to about 20 lines per day.

Does it make sense, then to program 4,6 or even 8h per day?

Would I really experience a drop of productivity if I did it for like 2h per day?

Maybe that is the wrong thing to ask.

A better thing to ask would be, would I really experience a drop in productivity if after doing a good chunk for the day, I finished it early?

Or rather, maybe the problem is that at some point my goals changed from maintaining desire to doing work for the simple reason that I mistrusted myself.

For me programming only really works after I boil my desire until I bursts and then let it out.

Maybe that process is what I should be refining towards rather than chaining myself with self imposed deadlines.

Very well. I'll do it like that.

For today I'll start now which is 4pm and go at it for as long as I feel like it.

First I will generalize the feedforward iteration. Then I will make the recurrent layers.

The narrower and more restricted the goal, the easier it is to manifest.