Let me commit here. I do not think I will continue past here.
Instead focus on doing work. What I need to focus on now that I have this is my desire to finish the char RNN. I will build it up and release it tomorrow.

All in all, even though it was not that long, I did write over 50 lines today. This is well over the 20 LOC per day average.

This is what I was talking about. 2h even started at 4pm can be a lot of work. On days when I am supposedly lazy and do do much work, I have plenty of stuff backed up in me.