I am up. Yesterday it looked like I would fall asleep, but I spend most of the night in intense concentration. Let me see if I got any replies and then I will move to the next step. Maybe I'll give the talk by Kurzweil a chance.
Also today I really need to remember to watch the Zesshoushinais.

Today is interesting. Even though there is snow outside, it is the warmest it has been in quite a while.

I really had a lot to think about last night, and it was mostly on what would come when I start doing RL.

Once I am done with the char RNN there is really quite a lot to come after that, a lot of it very difficult stuff that will absolutely push the library to its limit. I am really, really lucky that a convenient thing like poker is really my entry point into the field of RL. If it was something else there is a very real chance that I would not be able to handle myself. Both my own understanding and that of the broader field is just so low right now that the general purpose techniques might as well be worthless.

Even I had the net capable of perfectly modeling the game, for most games I could do jack with it. This is such a low level of control that I might as well be a baby. Because of that I won't be able to avoid putting my beliefs on the line as I go forward.

Let me finish with the char RNN thing and I will move to the next steps.

What I did with the iterator yesterday was very good work. That thing was not trivial at all to make.

Today what I need to do to make the char RNN work will be easier in comparison. But, to start things off let me watch the talk by Kurzweil. And I will change my approach as per pan. I will not use work as a yardstick anymore, but desire.