Considering how close I am to the goal, every part of the char-RNN is coming along with great difficulty.
It seems the more I do, the less motivated I feel to finish this. Honestly, it feels like this will take me another 3 weeks.

If so, fine. I'll deal with it to perfection and then take a break from all of this.

I have a very real problem right now. Poker might be on the boundary, but a lot of the very interesting things on RL are one level above the state of the art.

I put such amazing effort into this that I did not wonder too much how I am going to follow up once I actually have the library. This whole time I've been patching holes in my programming essentially.

I think subconciously I've been aware of this which why I am always pushing upwind against finish the library for the past few days.

https://www.reddit.com/r/programming/comments/80gxgl/classic_masterpiece_book_on_ai_and_lisp_now_free/dv2g2th/?context=3&st=je9yn57b&sh=d5a3a997

I had an interesting discussion with /u/combinatorylogic here.

I agree with him that nothing is on the horizon. Even I myself can only tell the direction.

It was always that way from the start.

It was that way when I started work on ML in late 2015, it was that way when I started work on Spiral and it is that way now.

Still, I do not feel up to the challenge of pushing the entire field of ML a level right now.

...
...

Actually, forget coding for today. Let me spend a few hours just planning what I will do when I am done.

I am leaning towards something very different from what I imagined originally, but the time has come to make a decision.