I could not quite gather my determination while I was thinking in bed. Maybe I'll manage it overnight, but here is what I am leaning towards.
Consider finishing the useless char-RNN a job and a leave it for the future. Complete it, document it and forget it for a year.

Somehow I found myself in a situation where each of the individual pieces in the library is greatly useful, but the actual end product the recurrent nets won't be like that for a while. I am not nearly as pessimistic as /u/combinatorylogic on this, and it won't be long before the dawn comes, but I am absolutely sick of the unpaid work.

Even if I finish the char RNN there is absolutely nothing I gain from having it. In the moment of truth, I am drawing a hard line on the necessity of being able to reason about the nets.

Since there is value in finishing the project and making its elements I am going to do it. I will treat it as a goal and scale the mountain just like I said I would.

After that I will come down from the mountain, look at what I have accomplished and get more realistic. It was not like this at all a few years ago, but making the Spiral language and its library has leveled me up insanely in terms of programming skill. Whereas some of the options weren't appealing back in 2016, I think that if I focus myself now, I can crush anybody in any game, neural nets or not by the sheer force of my programming ability. I have the Spiral power.

And since poker is on the boundary, I think that for the lower stakes, I could tackle it using classical techniques. I'll think of something. I think I could try using CFR with statistical techniques to come to a something.

RNNs are not the present, but they are certainly the future. No other algorithm has the capacity or the promise of them. I believe in them. It is absolutely inevitable that I will come back to them just like it was absolutely inevitable that I would come back to programming after ten years absence.

For the past 3 years, I just worked very, very hard.

It is not like me.

Here is what I will do. Since I worked like a mule for the past 3 years, for the next 3 how about I take a paid vacation? This is in contrast to the unpaid work that I did.

I want to change forms and go on the attack. I'll find some people online, kill them and take their stuff. I am just so hungry and my spirit is starved.

So my next projects should be the game library for poker and the interfaces. I could be done with all that plus the first agents using classical techniques in the next 3 months.

I sort of imagined that interfaces would be done using convolutional nets, but NNs really need to get about 3 order of magnitude easier to use for me to be able to do that. I wanted to absolutely do the architecture in a generic fashion, but Spiral and my own skills should suffice for that sort of thing now.

I already had the game library in F#, but I need to redo it in Spiral for ease of use. I'll also need some way of making GUIs.

Well, once I am done with that it will also be much easier to pick up the ML library again. Right now I have very little to use it on.

Let me call it a day here. Tomorrow I will just sit myself down, get rid of the distractions and begin work on finishing the damn net. It really won't be long for me to implement it, each of the optimizations, multiplicative integration and layer norm. Then I will update the tutorial. I guess I could be done with that in two weeks if I really apply myself, but I've been saying that for quite a while so we'll see.