I forgot to note something important.
It was really the thought experiment last night that convinced me.

I started by thinking about compressing the RNNs down to optimality. Once that is done, what would happen is that the model of a game for example would equal the game itself.

So in the end I'd have no way of using it for anything useful - for example, using MCTS on it would be equivalent to using it on the real game.

For a model to be useful it needs to be more than compressed. It needs to be better than the real thing.

I have absolutely no idea how to induce such a thing. Reward shaping won't cut it. Neither would expert iteration.

Well, combining MCTS and RNNs for poker might be viable since it is that sort of game, but I am not really in a hurry to try it out. I'll do it after I have proper baselines to compare it against.

Well, let me go back to resting and tomorrow I will resume work. There is only a month of it left anyway, so I'll put a good show.

After that will come play.