https://www.reddit.com/r/MachineLearning/comments/81jr1q/d_mit_agi_computational_universe_stephen_wolfram/?st=jeb2tt5l&sh=d4f8cfea
https://www.frontiersin.org/articles/10.3389/fncom.2017.00024/full

Yeah, I am not going to be programming today, or at least right now, not until I've come to a decision. I want to watch the lecture and read this paper and read Schmidhuber's paper in depth.

I've been doing more thinking last night and when I said I'd drop the library, I think that is just me losing my nerve.

I mean yes, I've been put on the spot with regards to solving intelligence. No, I have absolutely no idea how the brain communicates and yes, I pine for a way to automate it all.

But there is one ray of hope in doing this impossible task. That automatic algorithm that can continually evolve and optimize lower level algorithms is in fact me. I am the node from which everything starts.

I've always had only one approach - to scale my programming towards greater and greater abstraction. Abstraction is the one true friend that I have.

Even if I knew exactly what the brain did, probably, it would not be useful translated directly on the GPU. The best algorithm on meat is not the best on silicone. Rather I need to stick to my process.

Instead of trying to reason it out, I'd be better off doing good work and following the path of utility. I should just believe in myself. If I am following the right path in my journey to master games, the correct algorithm should come naturally to me. I've been acting fooling, the abstractions in the ML library right now are hardly the top of the mountain. It is in fact where the mountain starts! This is exactly the place to pull out everything, every gun that I have!

Of course, I am going to need some help to get over this hump so let me do some studying.

Maybe I'll work on the char RNN after that.