https://youtu.be/P7kX7BuHSFI?t=2827
Currently I am in middle of Wolfram's talk.

After I finish it, the plan is to take a look at this paper.

https://castlelab.princeton.edu/wp-content/uploads/2018/01/Powell-UnifiedFrameworkStochasticOptimization_Jan292018.pdf
A Unified Framework for Stochastic Optimization

I guess I'll spend the next few days like this until I gather my will and formalize my plan.

My plan A was to finish the library and the do ???

My plan B will be to finish the library and then use my abstraction skills to piece together the old and the new approaches to AI.

It will be a mess and a frankenstein moster of a thing, but I can at least hold out hope that along the way I will gain greater understanding.

Whenever I think of my ability to communicate with the net, I feel such a sense of inadequacy.

For the past year, I've been holding out hope that there would be a breakthrough in AI, but it seems the real benefit of the work that I did will be to allow me to make an even greater library than I originally envisioned.

I will have to do it. My own power of abstraction will have to be the lever to pry the door open.