https://arxiv.org/abs/1712.01815 Mastering Chess and Shogi by Self-Play with a General Reinforcement Learning Algorithm
Let me study the latest AlphaGo paper in depth. After that I will go through a course and go through Shmidhuber's paper's references.

My post on the ML sub got downvoted to 0 so I guess I have my answer there.

https://arxiv.org/abs/1611.05763
Learning to reinforcement learn

I already read this paper and surprisngly it is on bandits. This what I was also stuck on last year.

Yeah, I think that starting there might be fine.

I do not really need stochastic policies nor a scalar reward. Even in poker when the network needs to bluff in a balanced manner, it can just learn to do things like representing 7J as AA. Nanonoko does that and it seems to work for him.