https://www.youtube.com/watch?v=PTbxa6GsTWc&list=PLkFD6_40KJIznC9CDbVTjAF2oyt8_VAe3&index=3
Let me stop here.

It will take me a while to go through all of these lecture videos, but after that I will resume programming.

Getting up to the state of the art in RL is all I need in my studies here.

This is the same as the month of February last year before I mustered up the nerve to make Spiral. As long as I learn enough to gather orientation that should be fine.