Jesus, I am finally done.
Tomorrow I might skip a bunch of videos as I am not interested in optimal control. There will be a bunch of interesting stuff after that, but the last 5 lectures are guest lectures. Well, I guess I'll watch them.

At any rate, I am done for the day. It is not easy to go through 5 lectures a day.

I honestly, feel that I am wasting my time and that I should just skip to metalearning and watching the guest lectures.

I'll give it tomorrow and maybe Thursday and then I am going to get back into the game.