I am up. Let me start by going through the metalearning lecture. This means skipping the 10 inbetween for now. Then I will look at exploration. In fact, I will go through all of them after that.
I am forming my plan. Since I have 99% chance of failure, the only way to combat that and draw out that 1% would be to do a lot better science. I will make a powerful curriculum of progressively harder poker games. Doing it that way will absolutely tell me where the RNNs crash.

Yeah, I can accept that they wouldn't work right now, but they should work on the simplest of things. Once I find the exact boundary where they break, I will be able to work on extending it.

My view is that the RL researchers are too careless. Atari games are just too much to start things off. The reason they are getting such inconsistent results is because they are taking too large bites.

I do not know what two levels above the SOTA will be, but the first step to getting to one level above will be to recognize that unlike in the supervised learning regime, a NN architecture is not just about letting gradients flow.

There are hidden assumptions that haven't been touched on that need to be brought to light.

And I will be the lightbringer.