https://arxiv.org/pdf/1607.03474.pdf Recurrent Highway Networks
https://d4mucfpksywv.cloudfront.net/research-covers/reptile/reptile_update.pdf
Reptile- a Scalable Metalearning Algorithm

I've read over a dozen papers today in addition to finishing off the lectures, but the above two will be key.

If my suspicions are right Reptile will be the breakthrough that pushes RL into the real world and will be what will allow me to train viable agents for poker.

The RHN paper is the state of the art in LSTMs and will be what I will use for my agents. In 2016 I was enamored with MI and LN, but forget that for the time being.

The reason why the OpenAI post got upvoted so much is because others suspect what I do as well - that the reason RNNs train poorly is for the same reason feedforward nets do as well.

The question I've asked the author is intended to be rhetorical.

Tomorrow I will resume work on the char RNN. I will have to modify the way optimizer works, but nevermind that. The actually NN library is designed to be minimalist while being extremely flexible.

Let me read the paper again and then I will call it a day.

Damn the last week was quite something.

To think something like this would come out at this point in time - my luck is phenomenal. My 1% success rate literally went to 99% with this. Instead of working to succeed, what I will need to do is work not to fail.

Going forward from a position of strength is the ideal.

This breakthrough does not mean that all the problems are solved, not by a long shot, but at least for what I want to do, all the pieces are set.