The more I think about it, the more does the whole thing stink. You know let me just downgrade my expectations down to 1% and simply do this thing.
"Most of my confusion stems from the fact that in the Snail paper, it is directly benchmarked against Maml. This is strange because Snail is an NN architecture and Maml is an optimizer and the reasoning why this is done is never explained in the paper. This is a huge omission in my view because probably the reason why RNNs are hard to train on metalearning tasks is the same as for feedforward nets.

It is also unexplained what sort of architecture Maml is using, though I am guessing they took it from the original paper and run it themselves.

Considering that Snail is essentially a feedforward net with multiple input and output layers, it is never explained how it is able to do things which would require deep memory such as navigating a maze. I am left wondering what sequence length they could be using there.

Also on the test where they could not get Maml to train on the bandit problem, was that because they ran out of GPU memory (due to having to calculate higher order gradients) or did the net blow up?

I skimmed the survey, but it did not answer any of my questions."

I made this as a reply in the Reddit post. I guess I should have posted it on the top level, but I had to refine my complaints overnight. I only learned about Maml like 3 days ago.

Yeah, RL definitely feels like it would need a decade or two to get somewhere. Where does that leave my Singularity 2022 plans?

Well, I will know once I complete the char RNN and go into poker.

...

https://www.reddit.com/r/MachineLearning/comments/831wkk/r_reinforcement_learning_hidden_theory_and_new/?st=jejn5h5t&sh=b9520ae9

Now somebody posted this. Agh, let me just watch it and then I will start.