That was a waste of time. The level to cast Zap Q is definitely too high for me.
"The guy explains the algorithm in 1h of extremely dense, highly complicated math so I wouldn't be surprised about this. The paper follows a similar style and saying it is not the easiest paper to follow is really understating it. A PhD in stochastic approximation would be pretty much a prerequisite for understanding this.

A simplification would do much good here."

Let me see if I can finally resume programming before the commit section turns into a journal for whatever I am doing.

Focus, focus.

Ultimately, designing a NN to take advantage the structure of gambling games is a good idea that I should see to completion.

There are so many holes everywhere, and looking it more refinedly, an algorithm like Reptile could turn out to be anything from being the absolute key, to not doing anything.

I am a bit pissed that I cannot find anything on combining MAML with RNNs nor why SNAIL is being bechmarked directly against it as if it was a rival architecture instead of an optimizer.

But Reptile is simple enough that I'll be able to test it myself on various problems and it will serve as a proxy for Maml.

...Well, everything is good then.

Let me see where I was last time and then I will start by plugging a RNN in.