Ok, I finally fixed the split function.
That was a mess.

Things like this explain why progess is so slow.

Well, I am past it thankfully.

There is no need to dawdle anymore. Let me put in the recurrent net and I will try running it. No doubt another great debugging adventure is waiting for me, but I will have to do it.