So Karpathy used an LSTM to produce the samples. Well it might be possible to do it with an RNN, but I won't strain myself.
This is my first time training a sequence model.

This makes my goal for tomorrow clear. I will make a basic HRN and get it to perform well. I will bring in the softmax and when it does well on that I will checkpoint the repo and see just how fast I can make it go.

It should be a decent bit faster after I put all the optimizations in.

Edit: Forgot to push this an hour ago.