https://arxiv.org/abs/1611.01462 Tying Word Vectors and Word Classifiers: A Loss Framework for Language Modeling
https://arxiv.org/abs/1512.05287
A Theoretically Grounded Application of Dropout in Recurrent Neural Networks

https://arxiv.org/abs/1608.05859
Using the Output Embedding to Improve Language Models

So now I know how to do dropout and weight tying in RNNs. Well, I won't need weight tying. I thought it might have been convolution, but it turns out that it is merely tying the input and the ouput embedding layers in word models.

The variational dropout though might be good. It does make the net train longer though, so I'll skip it for the time being.

No doubt, it will come in handy in encouraging exploration in my poker players.

I've thought about exploration for a long time and essentially concluded that it is simply too difficult so what I am going to do is ignore the problem.

I am not that enamored with intrinsic curiousity, but I am absolutely going to use prediction as it will speed up training significantly.

For exploration I'll just let dropout take care of it.

In addition to that maybe I'll train it on a bunch of set lines, but that should be it. I won't bust my head over it.

Since there is no good way to deal with it, there is no need to think about it.

With that, let me finally start work on the HRN.

I'll start with a single layer one.