Enough. I am calling it a day here. I did well to find those bugs.
Yeah, this is it. This is more like the usual difficulty when it comes to dealing with neural nets. This is what I was afraid of. But this is not going to be something that will stop me. I am going to figure out what is eating the LSTM.

Programming is rarely a smooth journey. It is full of potholes like this.

I am thinking...

Well, my basic assumption here is that the gradients aren't being propagated correctly.

This seems hard to believe given that `add` and `hadmult` are so simple, but I am going to have to put them through the gradient checker.

Also...fuck, I lost the thought. There was another thing I wanted to do, but my just jettisoned that damn thing. It felt important too.

Ahhh....yes, I need to turn on the damn boundary checks. I remember forgetting to do this last time I had this kind of error and regretting it.

Well, instead of getting annoyed here I'll just stop for a bit. From what I can tell by looking at the outputs, the cost is being optimized in the right direction. Right now I do not feel like working on the gradient checker all over again. It will take me a few hours tomorrow.