It is time to start the hunt for this bug. There is not much left to go. Once I get the LSTM running I can put the rest of the optimizations in about a day.
I have two hypotheses for what is wrong:

1) I got the gradients for the backwards pass wrong somehow.

2) I am suspicious whether it will converge on a sequence of 1.

But really, I am pretty clueless right now. Bugs such as these always require a lot of creativty.

There is also the possibility that despite everything, I am still not doing the dataset correctly, but on a sequence of two it really should converge quickly.

So...I guess I'll start with the grad checker?