My plan is ruined. I thought I would just bring the LSTM in and it would work out of the box, but this is not training at all. The paper made it seem like it would work really well, but I this is not even training.
I spent a bunch of time on the gradient check, but it is a red herring. The reason why I am getting failures is definitely because the cost is huge so perturbing it has an outsized effect.

First, let me see if I can get it to converge on a single long sequence.