Done with the chores.
This short break dredged up my memories. I know what the problem is. I am certain of it. It is that sigmoid. Shit. Literally no RNNs have output into an activation function for a reason. That thing at the end should not be treated as a layer.

I did something really stupid here. Forget the GRU for a bit, let me test this out.