Yeah, that was it. i was correct. The HRN is unstable with cross entropy (I should clip the sigmoids), but with square error I can get it to memorize the sequence in 4 iterations.
Well, it does not go quite to zero, but it does go down from 4k to 64 and plateus.

Probably, there is a way to get it to zero, but that is not the purpose of this exercise. At least, now I can be sure that the library is working properly.

There were a bunch of bugs in the code, but the biggest one was in my understanding of how the nets should act.

As usual, my creativity saw me through.

Edit: Forgot to push this.

One thing that really bothers me right now is just how long the net takes to compile. Tomorrow I am going to put some effort into analysing exactly what is spiking the compilation times. I'll probably find a way to improve on them.

For now, let me see how this runs on a longer sequence.

https://arxiv.org/abs/1803.00590
Hierarchical Imitation and Reinforcement Learning

I'll also read this paper while I am it.

At any rate, this will be my last commit for the day. I won't do any more programming.