I am doing my best, but I am at the end of my rope here. I can't go on.
I managed to cut the code size down 5.1 to 4.3 and speed up compilation by about 20%.

This is not entirely satisfactory, but considering the amount of variables that are being tracked I can't expect much better than this.

I did a pretty good job.

That having said, there is a memory leak in this that I need to figure out. The reason why it worked up to now is because I was not clearing the right region.

What I have now makes sense to me, but does not work unfortunately. I need to figure it out.

There is something wrong here.

Today I really spent a lot of time reply to that guy. I am going to resist from interacting further on social sites as it ate up so much of my energy. I would be done with the memory bug by now had I started earlier.

Ah, well, let me rest a little. I am definitely going to deal with this tomorrow.