Ah...I get it. It is because I am calling region clear before triggering `bck`!
Ok, I figured it out.

How brilliant. Well, I guess with this I can really call it a day. I did what I set out to do which was investigate the code blowup and I even fixed a memory bug.

The generated code is definitely a lot nicer now, it doesn't have nearly as much useless unboxing as before. That is worth quite a lot to me.

But once thing that I understand now given how long it takes it to compile a single layer LSTM is the viability of the complex wavefront iteration that I decided no to do.

Yeah, I definitely couldn't stand a 10x increase in compilation footprint. That thing is absolutely non-viable unless I somehow figured out how to drastically speed up the compiler.

That is not going to be happening any time soon.

As it is, the compile times will put a practical limit on the kinds of models I can use.

Well, stacking layers won't be a problem as long as they are repetitive. I'll box them and put them in an array.

Now that I've come this far I need to figure out what comes next.

I think I'll plug in softmax and try sampling the sentences that come out tomorrow.

I am nearly there. With this I in essence have the ML library I've long been working towards.

A few steps more and I will have finished it all.