https://sites.google.com/view/hrlnips2017
The talks here honestly sound amazing. I really like the talk by David Silver that I just watched.

Right now I am agonizing just how long it takes it to compile an LSTM with just one layer. When I start stacking them and add RL on top just what the hell am I going to do.

Yeah, I know I said caching is the key, but I actually need the language to be faster at compiling. How do I make a language that does a lot of symbolic manipulation really fast?

What are my choices here exactly?

I'll put that aside. I'll complete the LSTM (well it already is), do the optimizations, write the tutorial and the pass it around for a bit and get feedback on how to improve it.

I should ask on the Ocaml sub on how to make this beast faster. A lot of people who write compilers and proves shouls have ideas for this.

I'll also start with the exercise of how to speed up the free variable function.

So for tomorrow, I think I will split my time between watching the talks and working on polishing the LSTMs.

...And polishing, I mean finally doing the softmax activation and cost and adding sampling to the mix.

It won't be too hard.

Apart from the huge issue of speeding up the compiler, I am now on the gentle path.

All the design problems with the library have essentially been worked out and starting from here I can look forward to reaping the rewards of my work.

Probably the next 3 years of my life will be more interesting than the last 3. I am looking forward to it.