Let me finally do some work. It turns out that more chapters came out than I expected so I'll save them for later.
Damn this let thing is just a whim. I did not really expect to be doing this right now.

But I am going to have to allocate some time for this now that performance is really an issue.

I guess I'll take a trip to the good old days of working on the language.

Forget about the LSTM. It will be done at some point. Hopefully once I am done with both it and the documentation I'll get some good advice from the Ocaml guys.