I am not sure, but the improvement does seem to be about 2%.
Ok, great.

I did well enough to put this in. 10% is nothing to scoff at. But it also shows that to get to the real deal, I am going to have to take more drastic measures.

There is no way I can be satisfied with taking 3s to compile a single LSTM layer. That is Theano tier compilation.

Well, no, Theano was still a lot slow, but I'd really prefer it if the language was 10-100x faster.

...

...

No I need to take this out. I am sort of relying on the filtering behavior for the sake of join points. Nevermind this.