Now it compiles in 1.1s with 5 layers.
This is it. In a burst of inspiration I've completely crushed the compile times down to reasonable.

It just goes to show that I barely know my own language.

My cost model of how expensive various compile time operations are is absolutely bonkers.

I've been treating them as free, but they are insanely expensive.