The test should be reasonably self contained now. The dataset is in the Temporary project folder. Apart from that the Cuda and VC++ dependecies will have to be installed by hand.
Regarding VC++'s old version being 3.7Gb, the VS installer is lying. It is actually much smaller than that.

The compiler is hardcoded to compile for 5.2 compute capability Cuda GPUs, but that can easily be adjusted and should probably be made automatic at some point.

This repo will go out of date soon enough, but by that time there will no doubt be challenging tests in the main repo.

Speed improvements and bug fixes to the compiler would definitely be appreciated

Good luck.