"3:45pm. https://www.reddit.com/r/MachineLearning/comments/84dwqf/d_r_hierarchical_rl_workshop_nips_2017_videos/?st=jer75umd&sh=8bfd1e98
Maybe I will get something here for why the talks won't load.

3:50pm. https://sites.google.com/view/hrlnips2017/papers?authuser=0

Well, thankfully there are papers for this."

Let me read the papers just for a bit, I do not feel like programming now. Maybe I'll skip on that for the rest of the day as I've been so distracted by things.

Since I can't really count that somebody will go over Spiral and look into how to make it faster, I'll put that out of mind for the time being and forget it exists.

Though it is very unlikely, it is possible that I've done a very good job. If a few other people could go over it and give me feedback that it can't be improved that would give me confidence in my implementation.

Though I am decently sure that the way hash consing could be done can be improved on since OCaml could do it much better.

Well, let me do a bit of studying as I plan what to do next.

I think I will give putting softmax in a shot.