No reply on the ML sub so far. I guess other people really do not have problems.
Now, I actually feel like doing programming for a bit now.

Really the trouble with softmax is that it behaves differently as an activation and as a cost function.

Then there is the whole thing with implementing sampling.

I will focus on the cost function for the time being. Let me put some thought into this.

I need to refresh my memory of how it works.