http://lambda-the-ultimate.org/node/2884
znmeb:
"1. One of the payoffs from bytecode (aka token threaded code) is cache and memory locality. ASTs tend to be spread out over the address space, making for more cache misses. The same can be said for other forms of threading. Ertl, Gregg and others make a pretty strong case for it.

2. The other key to efficient interpreters besides cache and memory location is making it easy for the processor to predict branch targets. I think that's also a consequence of doing threading right."

Leon P Smith:
"I would agree that bytecode interpreters have more opportunities for performance than syntax tree walking interpreters, and that they tend to be faster.

However, what I took the paragraph to mean was that the best time to perform optimizations is after the compiler converts a program to bytecodes, which isn't really true. Most compilers do get a lot of leverage out of such optimizations, however, they also tend to do quite a bit of work on various intermediate representations that are still pretty far removed from a linearized bytecode or machine-code format.

Once you've made it all the way to a mostly linearized representation, you've lost quite a bit of information about the program that makes it *much* more difficult to perform certain interesting optimizations, such as say, closure optimization, or trying to convert indirect calls into direct calls."

No, I am barking up the wrong tree here. I was wrong.

Partial evaluators really aren't interpreters.

Yeah, I could first translate the AST into bytecode, but the main essential problem with this is that the partial evaluation pass already does that!

Translation to bytecode is just linearization of the tree. Spiral does that naturally through let insertion.

So...

Then the main avenues to speed up Spiral would be this:

1) Improving compilation of patterns.

There are still places where this is inefficient. But this is trivial and I know how to do it.

2) Improving the data structures used.

The concestor dictionary that jdh mentioned could help. Also this area is a place of notable weakness for me. I should do some studying on immutable data structures. Maybe flay trees would be better than what I am using now? I remember them being noted for having good single threaded performance.

3) Improving the way hash consing is done.

Having to first unwrap the node is both an eyesore and an indirection. But if removed it, then it becomes confusing and I have no way of tracking which type is hash consed and which one is not.

Really, it all comes down to #3. Because I haven't done this perfectly, hash consing continues to act like an anchor on the compiler.

If I could relieve the pressure here, the rest of the optimizations could fall into place.

Otherwise, no matter what I do I will always will be swimming upstream.

Since OCaml is capable of tracking hash consing in the type system and F# isn't, it might be necessary to port the language to OCaml.

Let me take a break here and then I will look at the modular hash consing paper again.

It actually isn't important that I improve the performance right here, but that I come up with a plan for doing so.